Move ip-packet icmp error to `make` module
