+++ rust/connlib/tunnel/src/peer.rs
@@ -1,6 +1,6 @@
 use std::collections::{BTreeMap, BTreeSet, HashMap, HashSet, VecDeque, hash_map};
 use std::iter;
-use std::net::{IpAddr, Ipv4Addr, Ipv6Addr, SocketAddr};
+use std::net::{IpAddr, SocketAddr};
 use std::time::Instant;
 
 use crate::client::{IPV4_RESOURCES, IPV6_RESOURCES};
@@ -12,7 +12,7 @@ use dns_types::DomainName;
 use filter_engine::FilterEngine;
 use ip_network::{IpNetwork, Ipv4Network, Ipv6Network};
 use ip_network_table::IpNetworkTable;
-use ip_packet::{IpPacket, PacketBuilder, Protocol, UnsupportedProtocol, icmpv4, icmpv6};
+use ip_packet::{IpPacket, Protocol, UnsupportedProtocol};
 
 use crate::utils::network_contains_network;
 use crate::{GatewayEvent, IpConfig, otel};
@@ -294,13 +294,25 @@ impl ClientOnGateway {
         }
 
         let Some(state) = self.permanent_translations.get_mut(&packet.destination()) else {
-            return self.dst_unreachable(&packet);
+            return Ok(TranslateOutboundResult::DestinationUnreachable(
+                ip_packet::make::icmp_dst_unreachable(
+                    self.gateway_tun.v4,
+                    self.gateway_tun.v6,
+                    &packet,
+                )?,
+            ));
         };
 
         #[expect(clippy::collapsible_if, reason = "We want the feature flag separate.")]
         if firezone_telemetry::feature_flags::icmp_unreachable_instead_of_nat64() {
             if state.resolved_ip.is_ipv4() != dst.is_ipv4() {
-                return self.dst_unreachable(&packet);
+                return Ok(TranslateOutboundResult::DestinationUnreachable(
+                    ip_packet::make::icmp_dst_unreachable(
+                        self.gateway_tun.v4,
+                        self.gateway_tun.v6,
+                        &packet,
+                    )?,
+                ));
             }
         }
 
@@ -473,47 +485,6 @@ impl ClientOnGateway {
     pub fn id(&self) -> ClientId {
         self.id
     }
-
-    fn dst_unreachable(&self, packet: &IpPacket) -> Result<TranslateOutboundResult> {
-        let src = packet.source();
-
-        let icmp_error = match src {
-            IpAddr::V4(src) => icmpv4_network_unreachable(self.gateway_tun.v4, src, packet)?,
-            IpAddr::V6(src) => icmpv6_address_unreachable(self.gateway_tun.v6, src, packet)?,
-        };
-
-        Ok(TranslateOutboundResult::DestinationUnreachable(icmp_error))
-    }
-}
-
-fn icmpv4_network_unreachable(
-    src: Ipv4Addr,
-    dst: Ipv4Addr,
-    original_packet: &IpPacket,
-) -> Result<IpPacket, anyhow::Error> {
-    let builder = PacketBuilder::ipv4(src.octets(), dst.octets(), 20).icmpv4(
-        ip_packet::Icmpv4Type::DestinationUnreachable(icmpv4::DestUnreachableHeader::Network),
-    );
-    let payload = original_packet.packet();
-
-    let ip_packet = ip_packet::build!(builder, payload)?;
-
-    Ok(ip_packet)
-}
-
-fn icmpv6_address_unreachable(
-    src: Ipv6Addr,
-    dst: Ipv6Addr,
-    original_packet: &IpPacket,
-) -> Result<IpPacket, anyhow::Error> {
-    let builder = PacketBuilder::ipv6(src.octets(), dst.octets(), 20).icmpv6(
-        ip_packet::Icmpv6Type::DestinationUnreachable(icmpv6::DestUnreachableCode::Address),
-    );
-    let payload = original_packet.packet();
-
-    let ip_packet = ip_packet::build!(builder, payload)?;
-
-    Ok(ip_packet)
 }
 
 #[derive(Debug, PartialEq)]
